# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ContentType": {
        "enum": [
          "lesson",
          "problem",
          "transcript",
          "slide",
          "past_paper",
          "clo_doc",
          "student_note"
        ],
        "title": "ContentType",
        "type": "string"
      },
      "DeleteRequest": {
        "description": "XBLOCK_DELETED, or an orphan found by the reconciliation sweep (\u00a75.4).",
        "properties": {
          "offering_id": {
            "title": "Offering Id",
            "type": "string"
          },
          "tenant": {
            "title": "Tenant",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "usage_key": {
            "title": "Usage Key",
            "type": "string"
          }
        },
        "required": [
          "tenant",
          "offering_id",
          "usage_key",
          "trace_id"
        ],
        "title": "DeleteRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IngestAccepted": {
        "description": "What the service reports back after write-verify-swap (\u00a75.3).",
        "properties": {
          "blocks_received": {
            "title": "Blocks Received",
            "type": "integer"
          },
          "blocks_written": {
            "title": "Blocks Written",
            "type": "integer"
          },
          "chunks_written": {
            "title": "Chunks Written",
            "type": "integer"
          },
          "failed": {
            "items": {
              "type": "string"
            },
            "title": "Failed",
            "type": "array"
          },
          "swapped": {
            "title": "Swapped",
            "type": "boolean"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "blocks_received",
          "blocks_written",
          "chunks_written",
          "swapped"
        ],
        "title": "IngestAccepted",
        "type": "object"
      },
      "IngestRequest": {
        "description": "A batch of leaves for one course, from one bootstrap run or one publish.",
        "properties": {
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/LeafBlock"
            },
            "title": "Blocks",
            "type": "array"
          },
          "course_id": {
            "title": "Course Id",
            "type": "string"
          },
          "course_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Version"
          },
          "is_final": {
            "default": false,
            "title": "Is Final",
            "type": "boolean"
          },
          "offering_id": {
            "title": "Offering Id",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "tenant": {
            "title": "Tenant",
            "type": "string"
          },
          "topup": {
            "default": false,
            "title": "Topup",
            "type": "boolean"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "tenant",
          "course_id",
          "offering_id",
          "blocks",
          "trace_id",
          "run_id"
        ],
        "title": "IngestRequest",
        "type": "object"
      },
...
      "LeafBlock": {
        "description": "One published leaf block, extracted in-platform.",
        "properties": {
          "block_id": {
            "title": "Block Id",
            "type": "string"
          },
          "block_type": {
            "title": "Block Type",
            "type": "string"
          },
          "content_type": {
            "$ref": "#/components/schemas/ContentType"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "group_tokens": {
            "items": {
              "type": "string"
            },
            "title": "Group Tokens",
            "type": "array"
          },
          "publish_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publish Time"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "usage_key": {
            "title": "Usage Key",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "week": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Week"
          }
        },
        "required": [
          "usage_key",
          "block_id",
          "block_type",
          "version",
          "text",
          "content_type"
        ],
        "title": "LeafBlock",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "CourseMate",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/coursemate/api/ingest/blocks": {
      "post": {
        "description": "Chunk and index a batch of published leaf blocks.\n\n\u00a75.5 rule 1 \u2014 two blocks are never merged into one chunk \u2014 is enforced by the\nwire format rather than by a conditional here: each leaf arrives as its own\nrecord, so the chunker only ever sees one block's text and is structurally\nincapable of merging across them.",
        "operationId": "ingest_blocks_coursemate_api_ingest_blocks_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-coursemate-contract-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Coursemate-Contract-Version"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Blocks",
        "tags": [
          "ingest"
        ]
      }
    },
    "/coursemate/api/ingest/delete": {
      "post": {
        "description": "XBLOCK_DELETED, or an orphan found by the reconciliation sweep (\u00a75.4).",
        "operationId": "delete_blocks_coursemate_api_ingest_delete_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-coursemate-contract-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Coursemate-Contract-Version"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Blocks Coursemate Api Ingest Delete Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Blocks",
        "tags": [
          "ingest"
        ]
      }
    },
    "/coursemate/api/ingest/manifest/{offering_id}": {
      "get": {
        "description": "Every block currently served for this offering.\n\nThe reconciliation sweep (\u00a75.4) needs this because **no unpublish event\nexists in openedx-events.** Nothing tells us when an instructor unpublishes a\nunit, so the only way to detect it is to periodically compare what we serve\nagainst what the platform actually publishes. Without this the tutor keeps\nanswering from \u2014 and citing \u2014 content students can no longer see, which is a\ndirect Principle 3 violation that no event can catch.",
        "operationId": "manifest_coursemate_api_ingest_manifest__offering_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "offering_id",
            "required": true,
            "schema": {
              "title": "Offering Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-coursemate-contract-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Coursemate-Contract-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Manifest Coursemate Api Ingest Manifest  Offering Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Manifest",
        "tags": [
          "ingest"
        ]
      }
    },
    "/coursemate/api/ingest/offerings": {
      "get": {
        "description": "Courses the nightly sweep should visit.\n\nDeclared before /stats/{offering_id:path} would otherwise be reachable \u2014 a\npath-converter route is greedy, so ordering here is load-bearing.",
        "operationId": "offerings_coursemate_api_ingest_offerings_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-coursemate-contract-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Coursemate-Contract-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Offerings Coursemate Api Ingest Offerings Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Offerings",
        "tags": [
          "ingest"
        ]
      }
    },
    "/coursemate/api/ingest/prune": {
      "post": {
        "description": "Remove blocks the sweep found to be orphaned.\n\nSeparate from /delete, which drops a subtree on XBLOCK_DELETED. This takes an\nexplicit list the sweep computed, so a bug in tree-walking cannot cascade\ninto removing more than was asked for.",
        "operationId": "prune_coursemate_api_ingest_prune_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-coursemate-contract-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Coursemate-Contract-Version"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Request",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Prune Coursemate Api Ingest Prune Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prune",
        "tags": [
          "ingest"
        ]
      }
    },
    "/coursemate/api/ingest/stats/{offering_id}": {
      "get": {
        "operationId": "stats_coursemate_api_ingest_stats__offering_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "offering_id",
            "required": true,
            "schema": {
              "title": "Offering Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-coursemate-contract-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Coursemate-Contract-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Stats Coursemate Api Ingest Stats  Offering Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stats",
        "tags": [
          "ingest"
        ]
      }
    },
...
  }
}